글삭제 수정 권한확인(컨트롤러에서) 추가)

+++ src/main/java/com/cos/mybatisblog/controller/BoardController.java
@@ -28,7 +28,6 @@ public class BoardController {
 	@Autowired
 	private BoardRepository boardRepository;
 
-	
 	@GetMapping("")
 	public String postList(Model model) {
 		List<Board> boards = boardRepository.findAll(1);
@@ -36,20 +35,20 @@ public class BoardController {
 		// webapp/WEB-INF/views/post/list.jsp
 		return "redirect:/board/list/1";
 	}
-	
+
 	@GetMapping("/list/{page}")
 	public String boardList(Model model, @PathVariable int page) {
-		model.addAttribute("page",page);
+		model.addAttribute("page", page);
 		page = page - 1;
 		page = page * 3;
 		List<Board> boards = boardRepository.findAll(page);
 		Utils.setPreviewImg(boards); // 이미지를 previewImg에 저장
 		Utils.setPreviewContent(boards); // 이미지 태그 제거
-		model.addAttribute("boards", boards);		
+		model.addAttribute("boards", boards);
 		// webapp/WEB-INF/views/post/list.jsp
 		return "board/list";
 	}
-	
+
 	@GetMapping("/detail/{id}")
 	public String post(@PathVariable int id, Model model) {
 		try {
@@ -61,15 +60,15 @@ public class BoardController {
 		}
 		return "redirect:/";
 	}
-	
+
 	@GetMapping("/writeForm")
 	public String writeForm() {
 		return "board/writeForm";
 	}
-	
+
 	@PostMapping("/write")
-	public String save(Board board ,HttpSession session) {
-		User user = (User)session.getAttribute("user");
+	public String save(Board board, HttpSession session) {
+		User user = (User) session.getAttribute("user");
 		board.setUserId(user.getId());
 		board.setUsername(user.getUsername());
 		board.setUserProfile(user.getUserProfile());
@@ -81,27 +80,45 @@ public class BoardController {
 		}
 		return "redirect:/board";
 	}
-	
+
 	@PostMapping("/delete")
-	public @ResponseBody String detele(int id) {
+	public @ResponseBody String detele(int id, HttpSession session) {
+
+		User user = (User) session.getAttribute("user");
+		if (user.getId() != id) {
+			return Script.Back("권한이 없습니다.");
+		}
+
 		try {
 			boardRepository.delete(id);
 			return Script.alertAndHref("삭제되었습니다", "/board");
 		} catch (Exception e) {
 			e.printStackTrace();
 			return Script.Back("삭제 실패");
-		}		
+		}
 	}
-	
+
 	@GetMapping("/updateForm/{id}")
-	public String updateForm(@PathVariable int id, Model model) {
+	public String updateForm(@PathVariable int id, Model model, HttpSession session) {
+
+		User user = (User) session.getAttribute("user");
+		if (user.getId() != id) {
+			return Script.Back("권한이 없습니다.");
+		}
+
 		Board board = boardRepository.findById(id);
 		model.addAttribute("board", board);
 		return "board/updateForm";
 	}
-	
+
 	@PostMapping("/update")
-	public String update(Board board) { // param, form
+	public String update(Board board, HttpSession session) { // param, form
+
+		User user = (User) session.getAttribute("user");
+		if (user.getId() != board.getId()) {
+			return Script.Back("권한이 없습니다.");
+		}
+
 		try {
 			boardRepository.update(board);
 		} catch (Exception e) {
@@ -109,7 +126,5 @@ public class BoardController {
 		}
 		return "redirect:/board";
 	}
-
-	
 
 }
